Move VerificationError into its own module.
